:sparkles: Feat:미션통계 API 개발중

+++ back/seesaw/src/main/java/com/doacha/seesaw/model/dto/mission/CompareMissionResponse.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class CompareMissionResponse {
     private String missionId;
     private Double missionAverage;
-    private Double entireAverage;
+    private Long entireAverage;
 }
